+++ Spravy.Ui/Interfaces/ITaskProgressService.cs
@@ -2,6 +2,6 @@ namespace Spravy.Ui.Interfaces;
 
 public interface ITaskProgressService
 {
-    ConfiguredValueTaskAwaitable<Result<TaskProgressItem>> AddItemAsync(double impact);
+    ConfiguredValueTaskAwaitable<Result<TaskProgressItem>> AddItemAsync(ushort impact);
     ConfiguredValueTaskAwaitable<Result> DeleteItemAsync(TaskProgressItem item);
 }
